requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
firebase-admin==6.5.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.3.0
pyzbar==0.1.9          # optional — needs ZBar DLLs on Windows
python-barcode==0.15.1  # pure Python fallback
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
